pyproject.toml
[tool.poetry]
name = "plex-tools"
version = "1.0.0"
description = ""
authors = ["John Sturgeon <john@sturgeon.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
plexapi = "^4.15.7"
rich = "^13.7.1"
python-dotenv = "^1.0.1"
shazamio = "^0.6.0"
shazamapi = "^0.0.2"
thefuzz = "^0.22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
